Make the tests compatible with both Node and browser environment
