+++ packages/wix-ui-core/src/clients/GoogleMaps/IframeTestUtils.ts
@@ -3,18 +3,8 @@ const API_KEY_PARAM_NAME = 'key';
 
 export const getIframes = () => document.querySelectorAll('iframe');
 
-export const getIframeWithLangAndApiKey = (lang, apiKey) => {
-    const iframes = getIframes();
-    for (const iframe of Array.from(iframes)) {
-        const iframeScripts = iframe.contentWindow.document.querySelectorAll('script');
-        for (const iframeScript of Array.from(iframeScripts)) {
-            if (iframeScript.src.includes(`&${LANGUAGE_PARAM_NAME}=${lang}`) && iframeScript.src.includes(`${API_KEY_PARAM_NAME}=${apiKey}&`)) {
-                return iframe;
-            }
-        }
-    }
-    return undefined;
-};
+export const getIframeWithLangAndApiKey = (lang, apiKey): HTMLIFrameElement | null =>
+    document.querySelector(`iframe[data-lang=${lang}][data-api-key=${apiKey}]`);
 
 export const isIframeVisible = iframe => iframe.style.display !== 'none';
 
@@ -63,3 +53,14 @@ export class EventEmitterMock {
         this.eventListeners.forEach(callback => callback(event));
     }
 }
+
+export function IframesManagerMock () {
+    return;
+}
+
+IframesManagerMock.prototype = {
+    addIframe: jest.fn(),
+    getIframe: jest.fn(),
+    hasIframe: jest.fn(),
+    removeAllIframes: jest.fn(),
+};
