+++ packages/wix-ui-core/src/components/Badge/Badge.spec.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {badgeDriverFactory} from './Badge.driver';
 import {isEnzymeTestkitExists} from 'wix-ui-test-utils/enzyme';
-import {createDriverFactory} from 'wix-ui-test-utils/driver-factory';
+import {ReactDOMTestContainer} from '../../../test/dom-test-container';
 import {isTestkitExists} from 'wix-ui-test-utils/vanilla';
 import {badgeTestkitFactory} from '../../testkit';
 import {badgeTestkitFactory as enzymeBadgeTestkitFactory} from '../../testkit/enzyme';
@@ -9,8 +9,10 @@ import {mount} from 'enzyme';
 import {Badge} from './';
 
 describe('Badge', () => {
-
-  const createDriver = createDriverFactory(badgeDriverFactory);
+  const createDriver =
+    new ReactDOMTestContainer()
+    .unmountAfterEachTest()
+    .createLegacyRenderer(badgeDriverFactory);
 
   describe('children', () => {
     it('should be rendered', () => {
