+++ packages/wix-ui-core/src/components/StylableBadge/Badge.spec.tsx
@@ -1,16 +1,19 @@
 import * as React from 'react';
 import {badgeDriverFactory} from './Badge.driver';
 import {isEnzymeTestkitExists} from 'wix-ui-test-utils/enzyme';
-import {createDriverFactory} from 'wix-ui-test-utils/driver-factory';
+import {ReactDOMTestContainer} from '../../../test/dom-test-container';
 import {isTestkitExists} from 'wix-ui-test-utils/vanilla';
 import {stylableBadgeTestkitFactory as badgeTestkitFactory} from '../../testkit';
 import {stylableBadgeTestkitFactory as enzymeBadgeTestkitFactory} from '../../testkit/enzyme';
 import {mount} from 'enzyme';
 import {Badge} from './';
 
-describe('Badge', () => {
+describe('Stylable Badge', () => {
 
-  const createDriver = createDriverFactory(badgeDriverFactory);
+  const createDriver =
+    new ReactDOMTestContainer()
+    .unmountAfterEachTest()
+    .createLegacyRenderer(badgeDriverFactory);
 
   describe('children', () => {
     it('should be rendered', () => {
